Refactor platform support utils
